patched up mac build to work with 1.1.2

+++ RSDKv4/Audio.cpp
@@ -269,12 +269,12 @@ void ProcessMusicStream(Sint32 *stream, size_t bytes_wanted)
                     }
                 }
 
-                if (SDL_AudioStreamPut(musInfo.stream, musInfo.buffer, bytes_read) == -1)
+                if (SDL_AudioStreamPut(musInfo.stream, musInfo.buffer, (int)bytes_read) == -1)
                     return;
             }
 
             // Now that we know there are enough samples, read them and mix them
-            int bytes_done = SDL_AudioStreamGet(musInfo.stream, musInfo.buffer, bytes_wanted);
+            int bytes_done = SDL_AudioStreamGet(musInfo.stream, musInfo.buffer, (int)bytes_wanted);
             if (bytes_done == -1) {
                 return;
             }
@@ -403,7 +403,7 @@ void ProcessAudioPlayback(void *userdata, Uint8 *stream, int len)
                 samples = (unsigned long long)ov_pcm_total(&musInfo.vorbisFile, -1);
 
 #if RETRO_USING_SDL2
-                musInfo.stream = SDL_NewAudioStream(AUDIO_S16, musInfo.vorbisFile.vi->channels, musInfo.vorbisFile.vi->rate, audioDeviceFormat.format,
+                musInfo.stream = SDL_NewAudioStream(AUDIO_S16, musInfo.vorbisFile.vi->channels, (int)musInfo.vorbisFile.vi->rate, audioDeviceFormat.format,
                                                     audioDeviceFormat.channels, audioDeviceFormat.freq);
                 if (!musInfo.stream) {
                     printLog("Failed to create stream: %s", SDL_GetError());
@@ -485,7 +485,7 @@ void ProcessAudioPlayback(void *userdata, Uint8 *stream, int len)
                 }
 
 #if RETRO_USING_SDL1 || RETRO_USING_SDL2
-                ProcessAudioMixing(mix_buffer, buffer, samples_done, sfxVolume, sfx->pan);
+                ProcessAudioMixing(mix_buffer, buffer, (int)samples_done, sfxVolume, sfx->pan);
 #endif
             }
         }
