+++ RSDKv4/String.hpp
@@ -132,16 +132,15 @@ inline bool StrComp(const char *stringA, const char *stringB)
 
 inline int StrLength(const char *string)
 {
-    #ifdef USE_STDLIB
-    return strlen(string);
-
-    #else
+#ifdef USE_STDLIB
+    return (int)strlen(string);
+#else
     int len = 0;
     for (len = 0; string[len]; len++)
         ;
     return len;
 
-    #endif
+#endif
 }
 int FindStringToken(const char *string, const char *token, char stopID);
